Started re-structuring data_server and the load process to support additional data sources.
